@@ -97,6 +97,6 @@ public sealed partial class ReagentProducerAnomalyComponent : Component
     /// <summary>
     /// Solution where the substance is generated
     /// </summary>
-    [DataField("solutionRef")]
+    [ViewVariables]
     public Entity<SolutionComponent>? Solution = null;
 }
