@@ -25,7 +25,7 @@ public sealed partial class GasCondenserComponent : Component
     /// <summary>
     /// The solution that gases are condensed into.
     /// </summary>
-    [DataField]
+    [ViewVariables]
     public Entity<SolutionComponent>? Solution = null;
 
     /// <summary>
