@@ -16,7 +16,7 @@ public sealed partial class SmokeComponent : Component
     /// <summary>
     /// The solution on the entity with touch and ingestion reactions.
     /// </summary>
-    [DataField]
+    [ViewVariables]
     public Entity<SolutionComponent>? Solution = null;
 
     /// <summary>
