@@ -46,9 +46,9 @@ namespace Content.Shared.Mind
         ///     The first entity that this mind controlled. Used for round end information.
         ///     Might be relevant if the player has ghosted since.
         /// </summary>
-        [DataField, AutoNetworkedField]
-        public NetEntity? OriginalOwnedEntity;
-        // This is a net entity, because this field currently ddoes not get set to null when this entity is deleted.
+        [AutoNetworkedField]
+        public NetEntity? OriginalOwnedEntity; // TODO WeakEntityReference make this a Datafield again
+        // This is a net entity, because this field currently does not get set to null when this entity is deleted.
         // This is a lazy way to ensure that people check that the entity still exists.
         // TODO MIND Fix this properly by adding an OriginalMindContainerComponent or something like that.
 
